% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listr-package.R
\docType{package}
\name{LISTR-package}
\alias{LISTR}
\alias{LISTR-package}
\title{LISTR: Large-Scale Item-Response Statistics Tables}
\description{
Turns assessment and survey sample data (demographics, scores, weights, ability estimates and their IRT standard errors) into fully customizable pivot-style statistical tables. Provides weighted means, proportions above cut scores and proficiency-level percentages with measurement-error propagation from individual IRT standard errors, plus import from CSV, Excel, SPSS, SAS, Stata and IRT software person files, and formatted Excel export.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/weiandata/LISTR}
  \item Report bugs at \url{https://github.com/weiandata/LISTR/issues}
}

}
\author{
\strong{Maintainer}: Kunxiang Ma \email{makunxiang@weiandata.com}

Authors:
\itemize{
  \item Kunxiang Ma \email{makunxiang@weiandata.com}
}

Other contributors:
\itemize{
  \item WEIAN DATA TECH (Beijing) Co., Ltd. \email{contact@weiandata.com} [copyright holder, funder]
}

}
\keyword{internal}
