% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{lst_derive}
\alias{lst_derive}
\title{Add arbitrary derived variables (mutate-style)}
\usage{
lst_derive(x, ...)
}
\arguments{
\item{x}{A \code{listr_data} object.}

\item{...}{Name-value expressions evaluated in the data.}
}
\value{
\code{x} with derived columns added.
}
\description{
Add arbitrary derived variables (mutate-style)
}
