% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance-rep.R
\name{rep_factor}
\alias{rep_factor}
\title{Scale factor for replicate-weights variance}
\usage{
rep_factor(method, n_reps, fay_k = 0.5)
}
\arguments{
\item{method}{One of "fay", "brr", "jk1", "jk2".}

\item{n_reps}{Number of replicate weights R.}

\item{fay_k}{Fay factor (only for \code{method = "fay"}).}
}
\value{
Scalar variance factor.
}
\description{
fay: 1/(R*(1-k)^2) (PISA, k = 0.5); brr: 1/R;
jk1: (R-1)/R; jk2: 1 (TIMSS-style paired jackknife).
}
\keyword{internal}
