% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{st_count}
\alias{st_count}
\alias{st_wcount}
\title{Count and weighted-count statistics}
\usage{
st_count()

st_wcount()
}
\value{
A statistic spec for \code{\link[=lst_table]{lst_table()}}.
}
\description{
Count and weighted-count statistics
}
