% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{st_mean}
\alias{st_mean}
\title{Weighted mean statistic}
\usage{
st_mean(var)
}
\arguments{
\item{var}{Measure: a theta/score dimension name or a numeric column.}
}
\value{
A statistic spec for \code{\link[=lst_table]{lst_table()}}.
}
\description{
Weighted mean statistic
}
