% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var_measurement_mean}
\alias{var_measurement_mean}
\title{Measurement variance of a weighted mean of theta}
\usage{
var_measurement_mean(w, se)
}
\arguments{
\item{w}{Weights.}

\item{se}{Individual IRT standard errors.}
}
\value{
Measurement variance component.
}
\description{
Delta-method propagation of individual IRT standard errors:
\code{sum(w^2 * se^2) / sum(w)^2}.
}
\keyword{internal}
