% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var_measurement_prop}
\alias{var_measurement_prop}
\title{Measurement variance of a probabilistic proportion above a cutoff}
\usage{
var_measurement_prop(theta, se, cutoff, w)
}
\arguments{
\item{theta}{Ability estimates.}

\item{se}{Individual IRT standard errors.}

\item{cutoff}{Cut score.}

\item{w}{Weights.}
}
\value{
Measurement variance component.
}
\description{
\code{p_i = 1 - pnorm((c - theta_i)/se_i)};
\code{(dp/dtheta)^2 * se^2 = dnorm(z)^2}, hence
\code{sum(w^2 * dnorm(z)^2) / sum(w)^2}.
}
\keyword{internal}
