% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var_sampling_mean}
\alias{var_sampling_mean}
\title{Sampling variance of a weighted mean}
\usage{
var_sampling_mean(x, w)
}
\arguments{
\item{x}{Numeric values.}

\item{w}{Weights.}
}
\value{
Sampling variance of the weighted mean.
}
\description{
Linearized estimator: \code{sum(w^2 * (x - xbar)^2) / sum(w)^2}.
}
\keyword{internal}
