% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{lst_derive}
\alias{lst_derive}
\title{Add arbitrary derived variables (mutate-style)}
\usage{
lst_derive(x, ...)
}
\arguments{
\item{x}{A \code{listr_data} object.}

\item{...}{Name-value expressions evaluated in the data.}
}
\value{
\code{x} with derived columns added.
}
\description{
Add arbitrary derived variables (mutate-style)
}
\examples{
d <- data.frame(id = 1:5, part1 = 1:5, part2 = 6:10)
x <- lst_data(d, id = id)
x <- lst_derive(x, total = part1 + part2)
x$data$total
}
