% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{st_sd}
\alias{st_sd}
\title{Weighted standard deviation}
\usage{
st_sd(var)
}
\arguments{
\item{var}{Measure: a theta/score dimension name or a numeric column.}
}
\value{
A statistic spec for \code{\link[=lst_table]{lst_table()}}.
}
\description{
Weighted standard deviation
}
